(bug 42068) Exception on Incubator Special:Preferences
Visting [[incubator:Special:Preferences]] throws an exception
at the moment due to the unexpected altering of the
$wgDefaultUserOptions preference. I've changed it now so that
the preference is declared in the UserGetDefaultOptions hook
which is supposed for this.

Declaring this directly within WikimediaIncubator.php isn't possible
as this would prevent altering the $wmincPref variable.

Caused by https://gerrit.wikimedia.org/r/30554

(Untested)

Change-Id: I832c67808488656117b613d03a69ae9407fc84de