Fix "undefined index" error for optional "realtitle"
Note I'm also moving the str_replace( ' ', '_', … ) around.
Some of the places did this replacement already, but some
didn't. This appears to be a mistake. A title with a space
creates a possibly broken URL when used in a context that
doesn't expect URLs to contain spaces. I moved it to a place
where it can't be forgotten.

Bug: T268990
Change-Id: Ibd3127d45dd4466435e0ac4e36f760d2f1c3f17e