Don't apply title prefix filter in Special:Search for unset project
This hook applies a title prefix filter to the search engine, in the
case where the no project was set it was applying a prefix filter of
the string 'none', which resulted in no results returned to users.
Current fix is to not apply that exact string, although it seems there
is potential for other values that are also invalid.

It's not clear when this would have regressed, the code here and the
related search engine functionality have been constant for some time.

Bug: T266624
Change-Id: Ib9cd2df47b5e32c53f2735d8f417339a1fee7bd6