Change "None/All" to "All wikis"
The message is confusing to translators:
Why are there opposite things here? Is it "none" or "all"?

I guess that the intention was:
"NONE are specifically selected, and ALL are searched".

In context, the message refers to "all wikis", because
that's the effect of the action, so changing it
to say it explicitly.

Change-Id: I81f1c708eec01c2a70c10183939bb7c7593a8dd2