Add test wiki metadata to mw.config
If the page you are on is part of a test wiki, add metadata about
the test wiki to mw.config for use in JavaScript. The metadata that
is added is project, language code, prefix and real title; so if
the title is "Wp/xyz/Something", the prefix is "Wp/xyz", the project
is "p", the language code is "xyz" and the real title is "Something".

Bug: T307661
Change-Id: I0d1bdc6912be39c52035d54558c49f61a548adb2