Refactor and consolidate config
There were three different config options for essentially
the same things: $wmincProjects, $wmincSisterProjects and
$wmincProjectDatabases. This patch consolidates them all
into $wmincProjects, where each project (Wikipedia,
Wiktionary, etc.) is now an array of some metadata
that is needed in the extension.

Especially the changes to the getConf function will fix
a number of related bugs that first occured years ago
due to changes in how 'wikipedia' is handled in
InitialiseSettings.php; this patch should fix all of
them.

Bug: T117023
Bug: T126827
Bug: T130077
Bug: T162518
Bug: T307653
Bug: T307654
Change-Id: I9da823c164119b6f1d8dbe6e8d07b9b5ee1a2671