Add message group for on-wiki messages
Add a new subgroup of messages that are used on-wiki on the
Wikimedia Incubator. There are many advantages to having these
in Translatewiki proper instead of on-wiki localization,
not the least of which is the rate of translation (there
are more active translators on TWN than on the Incubator),
and better software support (translating individual messages
on-wiki with page translation is hacky).

Change-Id: I447ae0a86def433a3f46ef9f1017b13395e35d3a