Custom title handling
Encapsulate the test wiki prefix with <span> tags (with appropriate
classes) in order to allow separate styling of the prefix and the
actual title ("real title") of pages. Also encapsulate each of the
parts of the prefix in <span dir="ltr"> tags, so that they will
display in a more logical order for RTL languages.

Before this, the language converter would convert the entire title,
including prefix, resulting in (even more) nonsensical titles like
"Уқ/кк/Ана" for the Kazakh Wikiquote's page "Wq/kk/Ана" when
converted into Cyrillic. This change makes sure the language
converter only converts the "real title" part of the title.

Also remove the prefix entirely from the HTML <title> tag for
pages in the main namespace – since the prefix comes first, that
was the only bit of the title you would see if you had more than
a few tabs open at a time, and that's not very helpful.

Bug: T317098
Bug: T332427
Bug: T332452
Change-Id: I48a21d656ec0bdb3db7fb7158aea572a18bf5b5e