Add more on-wiki messages
Add messages that will be used in the ubiquitous {{test wiki}}
template on the Wikimedia Incubator. Half of these messages are
currently only localizable in the Wikimedia Incubator, and the
other half are not localizable at all.

Also change the messages for project names in English from an
unnatural "project language" order to "language project" order.

Change-Id: Ib76e0cf038fbf52ed56b2ce490ab252935f6d416