Remove {{#infopage:}} parser function
This parser function was added way back in 2011, but never
completed. There is no current need for a dedicated parser
function for this (on-wiki templates work just fine), and
the feature is never actually used.

The only hits for a search for "insource:/\#infopage:/" in
the Wikimedia Incubator are items on old software news pages
mentioning the beginning of development of this feature. It
is not used as an actual parser function anywhere.

Change-Id: If573a326da92951b07142389bca733c1cae382f9